renames inbox to inbox manager and adds inbox decodable
